requirements.txt
alembic==0.7.6
Babel==1.3
blinker==1.3
chardet==2.3.0
coverage==3.7.1
decorator==3.4.2
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-OpenID==1.2.4
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.11
flup==1.0.2
guess-language==0.2
itsdangerous==0.24
Jinja2==2.7.3
lamson==1.3.4
Mako==1.0.1
MarkupSafe==0.23
pbr==0.11.0
PyMySQL==0.6.6
python-openid==2.2.5
pytz==2015.4
six==1.9.0
speaklater==1.3
SQLAlchemy==1.0.6
SQLAlchemy-Enum-Dict==0.1.2
sqlalchemy-migrate==0.9.6
sqlparse==0.1.15
Tempita==0.5.2
Werkzeug==0.10.4
Whoosh==2.7.0
WTForms==2.0.2
